feat(site): make animation driver toggle per demo
