+++ apps/site/data/docs/components/alert-dialog/1.0.0.mdx
@@ -10,7 +10,7 @@ package: alert-dialog
 
 <Description>Show an alert prompt in a dialog.</Description>
 
-<HeroContainer>
+<HeroContainer showAnimationDriverControl>
   <AlertDialogDemo />
 </HeroContainer>
