+++ apps/site/data/docs/components/dialog/1.0.0.mdx
@@ -10,7 +10,7 @@ package: dialog
 
 <Description>Show a modal with configurable layout and accessible actions.</Description>
 
-<HeroContainer>
+<HeroContainer showAnimationDriverControl>
   <DialogDemo />
 </HeroContainer>
