+++ apps/site/data/docs/components/popover/1.0.0.mdx
@@ -10,7 +10,7 @@ package: popover
 
 <Description>Show content with a trigger in a floating pane.</Description>
 
-<HeroContainer>
+<HeroContainer showAnimationDriverControl>
   <PopoverDemo />
 </HeroContainer>
