+++ apps/site/data/docs/components/progress/1.0.0.mdx
@@ -10,7 +10,7 @@ package: progress
 
 <Description>Show percent completion with a progress bar.</Description>
 
-<HeroContainer>
+<HeroContainer showAnimationDriverControl>
   <ProgressDemo />
 </HeroContainer>
