+++ apps/site/data/docs/components/sheet/1.0.0.mdx
@@ -10,7 +10,7 @@ package: sheet
 
 <Description>A bottom sheet that slides up.</Description>
 
-<HeroContainer>
+<HeroContainer showAnimationDriverControl>
   <SheetDemo />
 </HeroContainer>
