+++ apps/site/data/docs/components/switch/1.0.0.mdx
@@ -10,7 +10,7 @@ package: switch
 
 <Description>Use in forms to toggle between two states.</Description>
 
-<HeroContainer>
+<HeroContainer showAnimationDriverControl>
   <SwitchDemo />
 </HeroContainer>
