+++ apps/site/data/docs/components/tabs/1.7.0.mdx
@@ -10,7 +10,7 @@ package: tabs
 
 <Description>Use in pages to manage sub-pages.</Description>
 
-<HeroContainer>
+<HeroContainer showAnimationDriverControl>
   <TabsDemo />
 </HeroContainer>
