+++ apps/site/data/docs/components/toast/1.9.1.mdx
@@ -10,7 +10,7 @@ package: toast
 
 <Description>Use to show feedback to user interactions</Description>
 
-<HeroContainer>
+<HeroContainer showAnimationDriverControl>
   <ToastDemo />
 </HeroContainer>
