+++ apps/site/data/docs/components/tooltip/1.0.0.mdx
@@ -10,7 +10,7 @@ package: tooltip
 
 <Description>A tooltip on web, with only accessibility output on native.</Description>
 
-<HeroContainer>
+<HeroContainer showAnimationDriverControl>
   <TooltipDemo />
 </HeroContainer>
