+++ apps/site/pages/docs/components/[...slug].tsx
@@ -11,7 +11,6 @@ import React, { useEffect, useState } from 'react'
 import { DocsPage } from '../../../components/DocsPage'
 import type { Frontmatter } from '../../../frontmatter'
 import { listeners } from '../../../hooks/setTinted'
-import { AnimationDriverTogglerContextProvider } from 'hooks/useAnimationDriverToggler'
 
 type Doc = {
   frontmatter: Frontmatter
@@ -56,9 +55,7 @@ export default function DocComponentsPage({ frontmatter, code }: Doc) {
       )} */}
       <MDXProvider frontmatter={frontmatter}>
         <ThemeTint disable={!isTinted}>
-          <AnimationDriverTogglerContextProvider>
-            <Component components={components as any} />
-          </AnimationDriverTogglerContextProvider>
+          <Component components={components as any} />
         </ThemeTint>
       </MDXProvider>
       <QuickNav key={frontmatter.slug} />
